Removed local uploads, updated asset structure
