+++ apps/cms-payload/src/collections/assets/AudioFiles.ts
@@ -1,7 +1,10 @@
 import type { CollectionConfig } from 'payload'
 
-export const NotationFiles: CollectionConfig = {
-  slug: 'notation-files',
+export const AudioUpload: CollectionConfig = {
+  slug: 'audio-upload',
+  admin: {
+    group: 'Assets',
+  },
   access: {
     read: () => true,
   },
@@ -18,6 +21,6 @@ export const NotationFiles: CollectionConfig = {
     },
   ],
   upload: {
-    mimeTypes: ['.musicxml', '.mxl', '.xml'],
+    mimeTypes: ['audio/mp3', 'audio/m4a', 'audio/ogg', 'audio/wav'],
   },
 }
