+++ apps/cms-payload/src/collections/assets/Documents.ts
@@ -2,6 +2,9 @@ import type { CollectionConfig } from 'payload'
 
 export const Documents: CollectionConfig = {
   slug: 'documents',
+  admin: {
+    group: 'Assets',
+  },
   access: {
     read: () => true,
   },
