+++ apps/cms-payload/src/collections/assets/Media.ts
@@ -2,6 +2,9 @@ import type { CollectionConfig } from 'payload'
 
 export const Media: CollectionConfig = {
   slug: 'media',
+  admin: {
+    group: 'Assets',
+  },
   access: {
     read: () => true,
   },
